+++ src/components/EditableField/EditableField.types.ts
@@ -43,7 +43,7 @@ export type Validation = {
 }
 
 export type CommitData = {
-  cause: 'BLUR' | 'ENTER' | 'OPTION_SELECTION' | 'DELETE_ACTION'
+  cause: string
   operation: string
   item: Object
 }
